+++ src/ExternalSystems/PaymentSystem.java
@@ -1,8 +1,8 @@
 package src.ExternalSystems;
 public interface PaymentSystem{
 
-    public boolean processPayment(int numTickets, String eventTitle, String studentEmail, int studentPhone, String epEmail, double transactionAmount);
+    public boolean processPayment(int numTickets, String eventTitle, String studentEmail, long studentPhone, String epEmail, double transactionAmount);
 
-    public boolean processRefund(int numTickets, String eventTitle, String studentEmail, int studentPhone, String epEmail, double transactionAmount, String organiserMsg);
+    public boolean processRefund(int numTickets, String eventTitle, String studentEmail, long studentPhone, String epEmail, double transactionAmount, String organiserMsg);
 
 }
